Started Android TV ReferenceAppKotlin project
This is just a basic Android Studio project file to serve as the base
for the new Android TV ReferenceAppKotlin project. This project is going
to be the primary reference for Android TV once it's ready. It will
demonstrate a variety of integrations, playback with ExoPlayer, and
other features over time.

Change-Id: I11b5f1410bb8f62439c20eb8e100e2d000698713