Added script to generate JSON for the app
This is a simple Python3 script that takes a CSV and outputs a JSON
file. The CSV should have the first row with the values of name, video,
thumbnail, and background. The script automatically maps between those
values and the ones used in the app, but it only does minimal error
handling.

Basic usage:

./atv-csv-to-json.py input.csv output.json

We will expand this later to add support for generating the media
actions feed and probably add better error handling and documentation,
but it's fully functional for the CSV-to-JSON conversion.

Change-Id: Iefec4b3aef08618de5711e934274c06105e18dc1