Added the core of home screen channels
This adds all the core support for creating a default home screen
channel and adding content to it from the VideoRepository. Some pieces
are missing from the Video data model (captured in another ticket), so I
added a couple of placeholders where non-null values are required to
just reuse the videoUri for now.

Since videos do not yet have unique IDs and deep links, running the
worker can put some "funky" data in the database, so the worker is not
yet tied to a broadcast receiver. It can be triggered manually for
testing though, which is what I did. Once we have the Video data model
updated and most of the content is real, we can add code to trigger the
worker.

Change-Id: Ie293804e7c78adae59a24cbb461529b30aee59fa