Added initial support for deep links
This adds support for deep links to playback specifically. This is
implemented via a ViewModel specifically for DeepLinks, which makes this
easier to test and means it will be easy to transition this to other
places in the app if ever needed. Another option would have been to
allow PlaybackFragment and its ViewModel to support deep links directly,
but that would complicate future support for a detail page, if desired.

Change-Id: I80388a6ecad5fe7798acbba9cf0cd01ac3b5bb21