Move all tests to /test from /androidTests
The instrumentation tests were mocking out the Android framework so there was no value to be added by having them run as instrumentation tests.

Robolectric provides a shim for the framework so we can remove some of the mocking of the android framework since a default implementation is provided for us.

Bug: 161985326
Change-Id: I0bf8f2f56bec4e718c2854205ffa0cd74b5483a6