Updated libraries to latest versions
Updated each of the libraries to the latest versions in the gradle
config. I didn't see anything in the release notes that would impact the
app and all tests still pass (plus browse and playback work as
expected), so this should be fine.

I did notice we were including ConstraintLayout but no longer using it,
so I removed that for now. If we end up needing it for a new layout, we
can always add it back.

Change-Id: I22419e4de46a53dec9eab71a6a7bd88cbe90582b