Fix of error messages for accessibility
The error messages during log in were not being announced via TalkBack
because it was an arbitrary TextView content change. In this case, the
error prevents the user from continuing, so the error message needs to
be announced as soon as it is set. By designating the accessibility live
region to assertive, accessibility services know that the view is
important and changes to it should be announced.

Bug: 184398998
Change-Id: I910f57803a320979421a8afd3b3cfa71ca2af2ae