Refactored player and player state listeners
This fixes a few issues with playback and the listeners.

1) We were using LiveData for communicating with the listeners. The
problem is that LiveData marks observers as inactive before the
Fragment.onStop method is triggered. Observers that are inactive are
still registered, but they don't receive any events. That means stopping
the player because it was being destroyed in onStop didn't actually send
the Pause event to any of the observers even though the LiveData value
was updated. This change uses a simple List of PlayerStateListeners to
trigger directly. The view model keeps the non-UI listeners around until
it is cleared. The Fragment is responsible for adding and removing its
UI listener.

2) We were using a PlaybackGlue.PlayerCallback to get play, pause, and
end events. This is nice because it's player-agnostic but it created a
problems with inconsistently receiving events, receiving multiple events
(e.g., receiving multiple pause events while actively scrubbing), and
receiving both pause and end events at the end of a video. This switches
to using ExoPlayer's Player.EventListener for the updates, which
exhibits none of these problems.

3) This change also fixes resuming from the correct playback position.
The LoadPlaybackStateListener now listens to pause events to know about
position updates. If it has seen a pause event, then it will just use
the timestamp from that event the next time it is loaded. This handles
both backgrounding and foregrounding where the Fragment is the same as
well as configuration changes where the Fragment is destroyed and
recreated.

Bug: 183732736
Change-Id: If2ec3a64d63e9e4e01fc4afd09c6ca29874e9141