Fixed issue (#59) building the LeanbackShowcase sample
An outdated version of Gradle was causing Android Studio to error, so I
updated that to the latest version. With the latest version, Dagger
failed to build. I updated it to 2.21 to fix the issues and updated the
generics usage, but this is far from the latest version. There is a
significant amount of work if we want to update to the latest version
that should be considered separately from this.

This also fixes a crash in MusicPlaybackService from the Application
being referenced before it is available and another crash in playback
due to VideoContent not having an accessible CREATOR.

Note: There is still a crash in the LiveData example due to the dagger
setup, but at least this PR gets everything building and all other
sections working again.